requirements.txt
django>=3.1.2
djangorestframework==3.12.1
django-cors-headers==3.5.0
django-filter
psycopg2